refactor: reorganize user routes
