--- app/api/user/avatar/route.ts
@@ -1,2 +1,0 @@
-// Deprecated alias - use /api/profile/avatar instead
-export { GET, POST, DELETE } from '../../profile/avatar/route';
