+++ src/services/user/__tests__/api-user.service.test.ts
@@ -13,7 +13,7 @@ describe('ApiUserService', () => {
       json: () => Promise.resolve({ data: { id: 'u1' } })
     });
     const res = await service.getUserProfile('u1');
-    expect(global.fetch).toHaveBeenCalledWith('/api/user/profile', expect.any(Object));
+    expect(global.fetch).toHaveBeenCalledWith('/api/profile', expect.any(Object));
     expect(res).toEqual({ id: 'u1' });
   });
 
@@ -24,7 +24,7 @@ describe('ApiUserService', () => {
       json: () => Promise.resolve({ data: { theme: 'dark' } })
     });
     const res = await service.updateUserPreferences('u1', { theme: 'dark' } as any);
-    expect((global.fetch as any).mock.calls[0][0]).toBe('/api/user/settings');
+    expect((global.fetch as any).mock.calls[0][0]).toBe('/api/settings');
     expect(res).toEqual({ success: true, preferences: { theme: 'dark' } });
   });
 });
